feat: map is separated
