+++ src/components/MainViewContainer/MainViewContainer.js
@@ -1,28 +1,13 @@
-import L from "leaflet";
 import React, { useMemo, useState } from "react";
-import { MapContainer, Marker, Popup, TileLayer } from "react-leaflet";
 import "react-leaflet-fullscreen/dist/styles.css";
-import MarkerClusterGroup from "react-leaflet-markercluster";
-import { Block } from "../../lib/Block/Block";
 import centers from "../../lib/cityCenters";
-import { hospitalIcon, pharmacyIcon, vetIcon } from "../../lib/Icons";
 import { useFetch } from "../../utils/hooks";
-import { DragIcon } from "../DragIcon/DragIcon";
 import { Footer } from "../Footer/Footer";
-import { FullScreenIcon } from "../FullScreenIcon/FullScreenIcon";
 import { HeaderCombined } from "../Header/HeaderCombined";
 import { FILTER, SEARCH_AT } from "../Header/HeaderRow";
-import InfoCard from "../InfoCard/InfoCard";
 import ListPage from "../ListPage/ListPage";
+import { CENTER_LAT, CENTER_LNG, MapPage } from "../MapPage/MapPage";
 import styles from "./MainViewContainer.module.scss";
-
-const CENTER_LAT = 37.683664;
-const CENTER_LNG = 38.322966;
-const ZOOM = 6;
-const MIN_ZOOM = 7;
-
-const LEFT_TOP_BOUND = [34.325514, 28.939165];
-const RIGHT_BOTTOM_BOUND = [41.57364, 42.770324];
 
 const MainViewContaier = () => {
   const { data: fetchedData } = useFetch(
@@ -33,12 +18,7 @@ const MainViewContaier = () => {
     "https://eczaneapi.afetharita.com/api/cityWithDistricts"
   );
 
-  const [mapRef, setMapRef] = React.useState();
-  const [dragActive, setDragActive] = React.useState(true);
-
-  const toggleDragActive = () => {
-    setDragActive((active) => !active);
-  };
+  const [map, setMap] = useState();
 
   const [searchAt, setSearchAt] = useState(SEARCH_AT.HARITA);
   const [filter, setFilter] = useState(FILTER.HEPSI);
@@ -46,29 +26,15 @@ const MainViewContaier = () => {
   const [selectedCity, setSelectedCity] = useState(null);
   const [selectedDist, setSelectedDist] = useState(null);
 
-  const center = [CENTER_LAT, CENTER_LNG];
-
-  const setIconFn = (type) => {
-    switch (type) {
-      case FILTER.HASTANE:
-        return hospitalIcon;
-      case FILTER.ECZANE:
-        return pharmacyIcon;
-      case FILTER.VETERINER:
-        return vetIcon;
-      default:
-        return hospitalIcon;
-    }
-  };
   const handleLock = () => {
-    if (mapRef.dragging.enabled()) {
-      mapRef.dragging.disable();
-      mapRef.zoomControl.disable();
-      mapRef.scrollWheelZoom.disable();
+    if (map.dragging.enabled()) {
+      map.dragging.disable();
+      map.zoomControl.disable();
+      map.scrollWheelZoom.disable();
     } else {
-      mapRef.dragging.enable();
-      mapRef.zoomControl.enable();
-      mapRef.scrollWheelZoom.enable();
+      map.dragging.enable();
+      map.zoomControl.enable();
+      map.scrollWheelZoom.enable();
     }
   };
 
@@ -76,18 +42,17 @@ const MainViewContaier = () => {
     if (city == null) {
       setSelectedCity(null);
       setSelectedDist(null);
-      mapRef.flyTo([CENTER_LAT, CENTER_LNG], 7);
+      map.flyTo([CENTER_LAT, CENTER_LNG], 7);
       return;
     }
     const lat = centers[city.key]?.lat;
     const lng = centers[city.key]?.lng;
-    mapRef.flyTo([lat, lng], 12);
+    map.flyTo([lat, lng], 12);
     setSelectedCity(city.id);
     setSelectedDist(null);
   };
 
   const districtMap = useMemo(() => {
-    console.log("this runs");
     const theMap = new Map();
     const allDistricts = cityData?.data?.map((city) => city.districts).flat();
     allDistricts?.forEach((district) => {
@@ -106,7 +71,7 @@ const MainViewContaier = () => {
           <div></div>
         </div>
       </div>
-    ); //LOADBAR EKLE
+    );
   }
 
   const typeFilteredData = allData?.filter(
@@ -129,11 +94,6 @@ const MainViewContaier = () => {
 
   const hasVetData = allData?.some((item) => item.typeId === FILTER.VETERINER);
 
-  const onLockClick = () => {
-    handleLock();
-    toggleDragActive();
-  };
-
   return (
     <div className={styles.mainViewContainerPaper}>
       <HeaderCombined
@@ -143,52 +103,16 @@ const MainViewContaier = () => {
         setFilter={setFilter}
         searchBarVal={searchBarVal}
         setSearchbarVal={setSearchbarVal}
-        hasVetData={hasVetData}
+        hasVetData={!!hasVetData}
       />
 
       {searchAt === SEARCH_AT.HARITA && (
-        <Block zeroPaddingOnMobile>
-          <div className={styles.mainViewContainerMapContainer}>
-            <MapContainer
-              whenCreated={setMapRef}
-              className="hazir-map" //class adı kendinize göre ayarlayabilirsiniz isterseniz
-              center={center} //CENTER BILGINIZ NEREDE İSE ORAYA KOYUNUZ
-              zoom={ZOOM} //ZOOM NE KADAR YAKINDA OLMASINI
-              minZoom={MIN_ZOOM}
-              tap={L.Browser.safari && L.Browser.mobile}
-              maxBounds={[LEFT_TOP_BOUND, RIGHT_BOTTOM_BOUND]}
-            >
-              <DragIcon dragActive={dragActive} onLockClick={onLockClick} />
-              <FullScreenIcon />
-
-              <TileLayer //Bu kısımda değişikliğe gerek yok
-                url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
-                attribution='&copy; <a href="http://osm.org/copyright">OpenStreetMap</a> contributors'
-              />
-
-              <MarkerClusterGroup>
-                {searchFilteredData?.map((station) => {
-                  return (
-                    <Marker
-                      icon={setIconFn(station.typeId, station.subTypeId)}
-                      key={station.id} //key kısmını da kendi datanıza göre ayarlayın mydaya.id gibi
-                      position={[station.latitude, station.longitude]} //Kendi pozisyonunuzu ekleyin buraya stationı değiştirin mydata.adress.latitude mydata.adress.longitude gibi
-                    >
-                      <Popup>
-                        <InfoCard
-                          key={station.id}
-                          item={station}
-                          districtMap={districtMap}
-                          styleName="popup"
-                        />
-                      </Popup>
-                    </Marker>
-                  );
-                })}
-              </MarkerClusterGroup>
-            </MapContainer>
-          </div>
-        </Block>
+        <MapPage
+          searchFilteredData={searchFilteredData}
+          districtMap={districtMap}
+          setMap={setMap}
+          handleLock={handleLock}
+        />
       )}
       {searchAt === SEARCH_AT.LISTE && (
         <ListPage
