+++ src/components/MainViewContainer/MainViewContainer.module.scss
@@ -6,10 +6,3 @@
     margin-bottom: 3rem;
   }
 }
-
-.mainViewContainerMapContainer {
-  display: flex;
-  height: 500px;
-  border-radius: 17px;
-  position: relative;
-}
