+++ src/components/MapPage/MapPage.module.scss
@@ -1,0 +1,6 @@
+.mainViewContainerMapContainer {
+  display: flex;
+  height: 500px;
+  border-radius: 17px;
+  position: relative;
+}
